Fix gross encoding errors everywhere
I misunderstood how Airflow Operators are instantiated, and
tried to process a templated field before the template was
rendered.
This was working by coincidence, until I tried to clean it up...

In an Operator __init__ method, templates have not been rendered
yet. There is a pre_execute() hook which can be overridden, and
it is run *after* the templates have rendered.